# apps/high-steak-api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: High Steak API
  description: |
    REST API for the High Steak social steak-rating app.

    ## Authorization

    Most endpoints require `Authorization: Bearer <JWT>`. Each operation lists required permission scopes in **`x-required-scopes`**. The JWT must include those values in its `scopes` claim (see login/register token).

    **Role and scopes are never returned in JSON responses** — clients must read `roles` and `scopes` from the JWT payload only.
  version: 1.0.0
  contact:
    name: High Steak
servers:
  - url: http://localhost:8080/api
    description: Local development
...
paths:
...
  /users/search:
    get:
      tags: [Users]
      summary: Search users to follow
      operationId: searchUsers
      x-required-scopes:
        - users:discover
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          description: Search term matched against username and display name.
          schema:
            type: string
            minLength: 2
            maxLength: 100
          example: ribeye
      responses:
        '200':
          description: Matching user profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserPublicProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /users/{id}:
    get:
      tags: [Users]
      summary: Public user profile
      operationId: getUserProfile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPublicProfile'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/{id}/posts:
    get:
      tags: [Posts]
      summary: Posts by user (authenticated)
      description: |
        Lists posts visible to the authenticated viewer on this profile. Followers-only
        posts are included only when the viewer follows the author or is the author.
      operationId: getUserPosts
      x-required-scopes:
        - posts:read
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /users:
    get:
      tags: [Users]
      summary: List all users (admin)
      operationId: listUsers
      x-required-scopes:
        - users:read
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT from `POST /auth/login` or `POST /auth/register`.

        Decode the payload to read authorization data (never returned in JSON bodies):
        - `roles` — array with one role name (e.g. `USER`, `MODERATOR`, `ADMIN`)
        - `scopes` — permission strings checked by the API (see `x-required-scopes` on each operation)

        Known scopes:
        - `posts:read`, `posts:read:own`, `posts:write`, `posts:delete:own`, `posts:delete:any`, `posts:moderate`
        - `comments:write`
        - `users:discover`, `users:read`, `users:manage`
        - `subscriptions:read`, `subscriptions:write`
...
  schemas:
    Username:
      type: string
      minLength: 3
      maxLength: 50
      pattern: '^[a-zA-Z][a-zA-Z0-9_-]*$'
      description: |
        Unique login handle (`users.username`). Letters, digits, underscores, and hyphens only;
        must start with a letter (not a number).
      example: steak_lover
...
    Email:
      type: string
      format: email
      maxLength: 255
      description: Account email stored in `users.email` (VARCHAR 255).
      example: chef@example.com
...
    DisplayName:
      type: string
      minLength: 2
      maxLength: 100
      description: Public name shown on posts and profiles (`users.display_name`).
      example: Hossam Kandel
    PostTitle:
      type: string
      minLength: 1
      maxLength: 120
      description: Short headline for the steak post (`steak_posts.title`).
      example: Perfect medium-rare ribeye
    PostComment:
      type: string
      maxLength: 65535
      description: Optional long-form caption (`steak_posts.comment`, TEXT).
      example: 28-day dry-aged ribeye, reverse seared, finished with rosemary butter.
    RestaurantName:
      type: string
      maxLength: 120
      description: Optional venue name (`steak_posts.restaurant_name`).
      example: The Prime Cut
    RestaurantLocation:
      type: string
      maxLength: 255
      description: Optional city or address hint (`steak_posts.restaurant_location`).
      example: Austin, TX
    SteakRating:
      type: integer
      format: int32
      minimum: 1
      maximum: 5
      description: Overall score from 1 (poor) to 5 (exceptional).
      example: 5
    PostVisibility:
      type: string
      enum: [PUBLIC, FOLLOWERS_ONLY]
      default: PUBLIC
      description: |
        Who can see the post. `PUBLIC` appears in the everyone feed.
        `FOLLOWERS_ONLY` is visible to the author, their followers, and moderators only.
      example: PUBLIC
...
    UserSummary:
      type: object
      required: [id, username, email, displayName]
      description: |
        Public account profile. Does **not** include role or scopes — read those from the JWT `roles` and `scopes` claims.
      properties:
        id:
          type: string
          format: uuid
          description: User UUID.
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        username:
          $ref: '#/components/schemas/Username'
        email:
          $ref: '#/components/schemas/Email'
        displayName:
          $ref: '#/components/schemas/DisplayName'
        avatarUrl:
          type: string
          nullable: true
          maxLength: 512
          description: Relative URL to profile photo, or null if unset.
          example: /uploads/avatar-abc.jpg
...
    UserPublicProfile:
      type: object
      required: [id, username, displayName, postCount, subscribed]
      properties:
        id:
          type: string
          format: uuid
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        username:
          $ref: '#/components/schemas/Username'
        displayName:
          $ref: '#/components/schemas/DisplayName'
        avatarUrl:
          type: string
          nullable: true
          maxLength: 512
          example: /uploads/avatar-abc.jpg
        postCount:
          type: integer
          format: int64
          minimum: 0
          description: Count of public, non-hidden posts.
          example: 12
        subscribed:
          type: boolean
          description: Whether the authenticated viewer follows this user.
          example: true
...
    AuthorSummary:
      type: object
      required: [id, displayName]
      properties:
        id:
          type: string
          format: uuid
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        displayName:
          $ref: '#/components/schemas/DisplayName'
    PostResponse:
      type: object
      required: [id, title, rating, imageUrls, createdAt, hidden, visibility, author, tags]
      properties:
        id:
          type: string
          format: uuid
          example: b2c3d4e5-f6a7-8901-bcde-f12345678901
        title:
          $ref: '#/components/schemas/PostTitle'
        comment:
          allOf:
            - $ref: '#/components/schemas/PostComment'
          nullable: true
        rating:
          $ref: '#/components/schemas/SteakRating'
        imageUrls:
          type: array
          minItems: 1
          items:
            type: string
            maxLength: 512
          description: Ordered photo URLs for the post.
          example: [/uploads/abc123.jpg, /uploads/def456.jpg]
        restaurantName:
          allOf:
            - $ref: '#/components/schemas/RestaurantName'
          nullable: true
        restaurantLocation:
          allOf:
            - $ref: '#/components/schemas/RestaurantLocation'
          nullable: true
        createdAt:
          type: string
          format: date-time
          description: When the post was published (ISO-8601 UTC).
          example: '2026-05-30T18:30:00Z'
        hidden:
          type: boolean
          description: True if moderators hid the post from public feeds.
          example: false
        visibility:
          $ref: '#/components/schemas/PostVisibility'
        author:
          $ref: '#/components/schemas/AuthorSummary'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/ReviewTagSummary'
    ReviewTagSummary:
      type: object
      required: [id, label, sentiment]
      properties:
        id:
          type: string
          format: uuid
          example: c3d4e5f6-a7b8-9012-cdef-123456789012
        label:
          type: string
          maxLength: 80
          description: Human-readable tag label.
          example: Juicy
        sentiment:
          type: string
          enum: [POSITIVE, NEGATIVE]
          example: POSITIVE
...
    ApiError:
      type: object
      required: [timestamp, status, error, message]
      properties:
        timestamp:
          type: string
          format: date-time
          example: '2026-05-30T19:00:00Z'
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Title must be at most 120 characters
        fieldErrors:
          type: object
          additionalProperties:
            type: string
          nullable: true
          example:
            title: must not be blank
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'